requirements.txt
pdfplumber==0.11.0
pypdf==4.2.0
reportlab==4.2.0
flask==3.0.3
streamlit==1.35.0
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.2
